fix : 수익률 계산 반환값 Long으로 변경

+++ src/main/java/lotto/service/YieldCalculateService.java
@@ -20,7 +20,7 @@ public class YieldCalculateService {
     }
 
     public void calculateYield() {
-        int totalPrizeMoney = winningLottos.getTotalWinningPrizeMoney();
+        long totalPrizeMoney = winningLottos.getTotalWinningPrizeMoney();
         wallet.calculateRateOfReturn(totalPrizeMoney);
     }
